firmware: scmi_agent: add SMCCC calls as transport layer
Implement SMCCC based transport for SCMI messages.

This change is related to the smc.c transport driver proposed by
Peng Fan [1] in the Linux kernel. It allows a SCMI agent to communicate
with a SCMI server hosted in the secure world of the CPU.

SCMI SMC transport driver is based on SMCCC v1.0 compliant calls.
The function ID used by the agent to identify itself towards the
SCMI server is set by the 32bit property "arm,smc-id" in the U-Boot DT.

Links: [1] https://lkml.org/lkml/2020/3/8/141
Change-Id: I3f984b644583cb7080d779898cebfed8669d3f33
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>